fixes and improvements from review * move header components to components folder * change toggle/link colors to match design * revert/uncomment ally test
