+++ frontend/src/components/Agreements/AgreementBudgetLinesHeader.js
@@ -24,11 +24,10 @@ export const AgreementDetailHeader = ({ heading, details, includeDrafts, setIncl
                     <FontAwesomeIcon
                         icon={includeDrafts ? faToggleOn : faToggleOff}
                         size="xl"
-                        className="margin-right-1 cursor-pointer usa-tooltip"
+                        className={`margin-right-1 cursor-pointer ${includeDrafts ? "text-primary" : "text-base"}`}
                         title={includeDrafts ? "On (Drafts included)" : "Off (Drafts excluded)"}
-                        // data-position="top"
                     />
-                    <span className="text-primary">Include Drafts</span>
+                    <span className="text-ink">Include Drafts</span>
                 </button>
             </div>
             {details && <p className="font-sans-sm">{details}</p>}
